Refactor: split monolithic CanvasArea into modular components and hooks
